# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an auto-generated API Docs.",
        "title": "Roast GitHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "jakueenak@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/roast/github": {
            "post": {
                "description": "Get roast by username and data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roast"
                ],
                "summary": "get roast by username and data",
                "parameters": [
                    {
                        "description": "Prompt",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GithubRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GithubContentResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ContentResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ContentResponseFailure"
                        }
                    }
                }
            }
        },
        "/roast/resume": {
            "post": {
                "description": "Get roast by resume.pdf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roast"
                ],
                "summary": "get roast by resume.pdf",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Resume as PDF",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language for the content default is id",
                        "name": "lang",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Gemini API key",
                        "name": "key",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ContentResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ContentResponseFailure"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ContentResponseFailure"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ContentResponse": {
            "type": "object",
            "properties": {
                "generated_content": {
                    "type": "string"
                }
            }
        },
        "models.ContentResponseFailure": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.ContentResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.ContentResponse"
                }
            }
        },
        "models.GithubContentResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "generated_content": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.GithubContentResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.GithubContentResponse"
                }
            }
        },
        "models.GithubRequest": {
            "type": "object",
            "required": [
                "lang",
                "username"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "lang": {
                    "enum": [
                        "auto",
                        "id",
                        "en"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Language"
                        }
                    ]
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                }
            }
        },
        "models.Language": {
            "type": "string",
            "enum": [
                "auto",
                "id",
                "en"
            ],
            "x-enum-varnames": [
                "LangAuto",
                "LangID",
                "LangEN"
            ]
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}